Pi-stacking fix (#97)
The Pi stacking interactions have been changed for a more accurate implementation: it now relies on the angle between the vector normal to a ring's plane and the vector between centroids (`normal_centroids_angles`) instead of the `shortest_distance` parameter.  In addition, the `EdgeToFace` interaction makes sure that the intersection between both ring planes falls inside one of the rings.